pass constructor.New() a copy of buffer instead
According to doc on Napi::Buffer::New(napi_env,T*,size_t):
The Napi::Buffer object does not assume ownership for the data and
expects it to be valid for the lifetime of the object.

Since the Napi::Buffer is subject to garbage collection this overload
is only suitable for data which is static and never needs to be freed.

So use Buffer::Copy() instead of Buffer::New() to pass the VSAM's
layout data to the new constructor.