remove unneeded freopen, refactor file exist checks
To check if file exits, code unnecessarily fopen()'d
with rb+ instead of rb; only changing to rb requires
checking the case when file exists but empty and opened
in read-only mode, in which case fopen would also fail
(VSAM restriction).

Same for Exist(), which opened the file with rb+.
Now the file is fopen()'d with rb but with error checking
to process the above case.

Also VsamFile::lastrc_ variable was being set to -1 on error
in several places while still left it as 0 in some cases;
this is resolved in this commit.