test: add test that hangs
The find() call in this test hangs because the write() was
performed from a different thread than the thread that opened
the VSAM dataset, and the write() resulted in the file being
extended.

This is documented here:
https://www.ibm.com/support/knowledgecenter/SSLTBW_2.3.0/com.ibm.zos.v2r3.cbcpx01/cbc1p2274.htm

i.e.:
"MVS files that are opened by data-set names or ddnames are
thread-specific in the following ways:

MVS files opened in update mode where repositioning functions are
used after the files have been extended are also restricted to the
owning thread. This is because a reposition might need to reopen
the read DCB in order to be able to see the new EOF marker. The
runtime library does not enforce this restriction.

Multivolume ..