src: route I/O to a single thread if open in rw mode
https://www.ibm.com/support/knowledgecenter/SSLTBW_2.3.0/com.ibm.zos.v2r3.cbcpx01/cbc1p2274.htm
states:

"MVS files that are opened by data-set names or ddnames are thread-specific
in the following ways:

MVS files opened in update mode where repositioning functions are used
after the files have been extended are also restricted to the owning thread.
..."

Since the vsam.js APIs (open, write, update, delete, find, close, etc.) can
occur from different threads, this commit creates a single thread per
openSync() or allocSync() call, only when the dataset is being opened in
read-write mode, and routes (in a FIFO queue) all I/O calls to that thread,
wait for the thread to execute it, and return the result to the caller,
often through the provided callback function.

When the file is opened in read-only mode (by passing "rb,type=record" to
openSync()), then no such thread is created, and the I/O are executed by
whichever thread made the API call.