src: handle undefined fields in write and update
Before this update an undefined field in a given record would contain
the text 'undefined' and would be written (write and update) as such
if that value was valid (string type and length > 8).

With this commit, for write: an undefined field is set to "" for string
type, and all 0's for hexadecimal type, unless the field's minLength > 0,
in which case an error is thrown. And for update, an error is thrown if
a field was undefined.