src: replace findUpdate & findDelete by update and delete
Now the number of records updated or deleted by the versions of update
and delete is passed back as the first argument to the callback function.
So findUpdate() and findDelete() are no longer necessary (their purpose
was to save having to send a record back if not needed, but now only a
count is passed back).