src: change default read-write open mode from ab+ to rb+
It seems the original ab+ mode was intended to create the file
if it didn't exist, but if it didn't exist, the dataset requires
that it's allocated first (so it must exist before openSync() is
called).

The 'ab+' mode also caused open to fail in read-write mode while
another process already has the dataset (even if in read-only mode),
with error:

EDC5061I An error occurred when attempting to define
a file to the system. (errno2=0xc00b0403).

With 'ab+' mode, the error doesn't occur if multiple read-only and one
read-write open is performed from the same process.