src: handle API calls made on a closed dataset
If user called an API, say close() or write(), while the dataset is
already closed, send an error to the callback or throw an error for
the Sync APIs, otherwise the user app will crash while accessing
a NULL pVsamFile pointer, unless built with -DDEBUG, in which case
the APIs would fail with an assertion error, except for close() which
didn't have that.

Also fix FindDelete_() to return the exit code from Find(), instead
of always return 0.